Moved api client test sources
